+++ database/migrations/2021_03_05_000001_create_access_tokens_table.php
@@ -4,16 +4,16 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreatePersonalAccessTokensTable extends Migration
+class CreateAccessTokensTable extends Migration
 {
     public function up()
     {
-        Schema::create('personal_access_tokens', function (Blueprint $table) {
+        Schema::create('access_tokens', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->morphs('tokenable');
-            $table->string('name');
             $table->string('token', 64)->unique();
             $table->text('abilities')->nullable();
+            $table->string('name')->nullable();
             $table->timestamp('last_used_at')->nullable();
             $table->timestamps();
         });
@@ -21,6 +21,6 @@ class CreatePersonalAccessTokensTable extends Migration
 
     public function down()
     {
-        Schema::dropIfExists('personal_access_tokens');
+        Schema::dropIfExists('access_tokens');
     }
 }