+++ src/Testing/Concerns/InteractsWithAuthentication.php
@@ -2,14 +2,14 @@
 
 namespace Butler\Service\Testing\Concerns;
 
+use Butler\Auth\ButlerAuth;
 use Butler\Service\Models\Consumer;
-use Laravel\Sanctum\Sanctum;
 
 trait InteractsWithAuthentication
 {
     public function actingAsConsumer(array $data = [], $abilities = ['*']): self
     {
-        Sanctum::actingAs(new Consumer($data), $abilities);
+        ButlerAuth::actingAs(new Consumer($data), $abilities);
 
         return $this;
     }